Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@7478 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ knuth/src/nu/mine/mosher/random/RNGDefault.java
@@ -10,12 +10,14 @@ package nu.mine.mosher.random;
  */
 public abstract class RNGDefault implements RandomNumberGenerator
 {
-	private final int seed;
+	private final long seed;
 
 	/**
+	 * Initializes the <code>RandomNumberGenerator</code> with
+	 * the given seed number.
 	 * @param seed
 	 */
-	public RNGDefault(final int seed)
+	public RNGDefault(final long seed)
 	{
 		this.seed = seed;
 	}
